Run home-assistant with host networking mode
It seems that it needs this for successful auto-discovery of devices.
I could manually work through a list of all ports required for all of
the types of devices I want to use, but this seems tedious.

Perhaps I'll move home-assistant to a separate host to prevent port
conflicts with any other services I want to run...