Configure tmux with some basic settings
Annoyingly, the home-manager tmux module loads the tmux-sensible plugin
and then sets a bunch of settings to the default values in tmux if
they're not specified in the nix module options.

This means that quite a few of the settings that the sensible plugin
sets to "better defaults" are effectively undone again by the nix
module, unless I explicitly set them again in my config.

Maybe I'll make a PR to home-manager for this one day...